+++ extension/repo/menu@a6621988-551d-495a-97d8-3c568bca2e9e/client.js
@@ -25,6 +25,7 @@ web.whenReady([sidebarSelector]).then(async () => {
       list: await fs.getJSON('https://notion-enhancer.github.io/notifications.json'),
       dismissed: await storage.get(_id, 'notifications', []),
     };
+  console.log($enhancerSidebarElement);
   notifications.waiting = notifications.list.filter(
     ({ id }) => !notifications.dismissed.includes(id)
   );
@@ -48,4 +49,4 @@ web.whenReady([sidebarSelector]).then(async () => {
   setTheme();
   document.querySelector(sidebarSelector).appendChild($enhancerSidebarElement);
 });
-web.hotkeyListener(await storage.get(_id, 'hotkey.focustoggle'), env.openEnhancerMenu);
+web.addHotkeyListener(await storage.get(_id, 'hotkey.focustoggle'), env.openEnhancerMenu);