+++ extension/repo/tweaks@5174a483-c88d-4bf8-a95f-35cd330b76e2/client.js
@@ -14,9 +14,8 @@ web.whenReady().then(async () => {
   if (cssInsert) {
     document.body.append(
       web.createElement(
-        web.html`<style id="${await storage.get(_id, 'insert.css')}@${_id}">
-        ${cssInsert}
-        </style>`
+        web.html`
+        <style id="${await storage.get(_id, 'insert.css')}@${_id}">${cssInsert}</style>`
       )
     );
   }